moved source table into a separate file. Will update group sources page in another commit
